# openapi.yaml
openapi: 3.0.3
info:
  title: Drone Delivery Management API
  version: 1.0.0
  description: API for the Penny Software technical assessment. 
servers:
  - url: http://localhost:8080/api/v1
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Location:
      type: object
      required: [lat, lon]
      properties:
        lat:
          type: number
          format: double
        lon:
          type: number
          format: double
        timestamp:
          type: string
          format: date-time
...
    Drone:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
          enum: [available, reserved, delivering, broken]
        location:
          $ref: '#/components/schemas/Location'
        current_order_id:
          type: string
          format: uuid
...
security:
  - BearerAuth: []
paths:
...
  /drones:
    get:
      summary: List drones (admin)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: list of drones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Drone'
  /drones/{droneId}:
    parameters:
      - in: path
        name: droneId
        required: true
        schema:
          type: string

    get:
      summary: Get drone details
      security:
        - BearerAuth: []
      responses:
        '200':
          description: drone
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Drone'
  /drones/{droneId}/location:
    post:
      summary: Drone posts its current location (heartbeat)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Location'
      responses:
        '200':
          description: heartbeat accepted
...